Joint estimation of the HT model parameters

Joint estimation of the HT model parameters assuming that the parameters for two different stability assays are the same. For a description of the arguments see the 'negloglikeHT.R' documentation

pairwise.fitHT(pars, dc1, dc2, r1, r2, gens, NCols1, NCols2)

Arguments

pars
a vector of the initial fraction, the segregation, the plasmid cost, the HT constant and the HT probability. All the parameters, except for the cost, have to be specified in logit scale (remember that logit 'p' is logit(p) = log(p)-log(1-p)). The cost is given in the log scale

dc1
is a matrix with the days and counts of the first data set. The first column are the days, the next columns are the counts over time of the plasmid-free colonies. One column per replicated platings.

dc2
is a matrix with the days and counts of the second data set. The first column are the days, the next columns are the counts over time of the plasmid-free colonies. One column per replicated platings.

r1
the number of replicated platings per day, for data set 1

r2
the number of replicated platings per day, for data set 2

gens
the number of generations per day

NCols1
A matrix whose number of rows is the number of days that the first assay was ran. The number of columns corresponds to the number of replicated platings. The entries correspond to the total number of colonies picked for testing in a particular day and replicate.

NCols2
Same as NCols1, but for the second assay's data (dc2)